i have finished most of the assignment
