+++ index.html
@@ -5,8 +5,8 @@
     <title>Lambda School Newsfeed</title>
     <link href="https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,400;0,700;1,400&display=swap" rel="stylesheet">
     <link rel="stylesheet" href="./styles/index.less">
-    <script src="./components/Menu.js" defer></script>
-    <script src="./components/Article.js" defer></script>
+    <!-- <script src="./src/components/Menu.js" defer></script> -->
+    <script src="./src/components/Article.js" defer></script>
   </head>
   <body>
     <div class="header">
