+++ package.json
@@ -19,5 +19,9 @@
     "eslint": "^7.6.0",
     "less": "^3.12.2",
     "parcel-bundler": "^1.12.4"
+  },
+  "dependencies": {
+    "axios": "^0.21.1",
+    "styled-components": "^5.3.0"
   }
 }
